+++ server/nearby/src/main/java/com/my/nearby/area/controller/SidoController.java
@@ -1,5 +1,6 @@
-package com.my.nearby.area;
+package com.my.nearby.area.controller;
 
+import com.my.nearby.area.model.service.SidoService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -7,12 +8,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/gugun")
+@RequestMapping("/sido")
 @RequiredArgsConstructor
-public class GugunController {
-    private final GugunService gugunService;
+public class SidoController {
+    private final SidoService sidoService;
     @GetMapping
-    public ResponseEntity<?> getGugunList() {
-        return ResponseEntity.ok(gugunService.findAllGugun());
+    public ResponseEntity<?> getSidoList() {
+        return ResponseEntity.ok(sidoService.findAllSido());
     }
 }
