+++ server/nearby/src/main/java/com/my/nearby/area/model/dto/GugunResponseDto.java
@@ -1,6 +1,7 @@
-package com.my.nearby.area;
+package com.my.nearby.area.model.dto;
 
 import com.my.nearby.BusinessException;
+import com.my.nearby.area.model.vo.GugunVo;
 import org.springframework.http.HttpStatus;
 
 public record GugunResponseDto(
